Move Time to shared_data_struct
import fix for Time

fixing minecraft_specs import
