+++ agents/craftassist/tests/fake_agent.py
@@ -13,7 +13,7 @@ from droidlet.shared_data_struct.base_util import TICKS_PER_SEC
 from droidlet.memory.craftassist.mc_memory import MCAgentMemory
 from droidlet.memory.craftassist.mc_memory_nodes import VoxelObjectNode
 from agents.craftassist.craftassist_agent import CraftAssistAgent
-from droidlet.shared_data_struct.base_util import Time
+from droidlet.shared_data_struct.shared_data_structs import Time
 from droidlet.dialog.dialogue_manager import DialogueManager
 from droidlet.dialog.droidlet_nsp_model_wrapper import DroidletNSPModelWrapper
 from droidlet.dialog.craftassist.dialogue_objects import (
