+++ droidlet/shared_data_struct/shared_data_structs.py
@@ -1,5 +1,7 @@
+import time
 from droidlet.interpreter.condition import AlwaysCondition, NeverCondition, NotCondition, TaskStatusCondition
 from droidlet.memory.memory_nodes import TaskNode, TripleNode
+from droidlet.shared_data_struct.base_util import TICKS_PER_SEC
 
 
 class Task(object):
@@ -123,3 +125,22 @@ def get_default_conditions(task_data, agent, task):
 
     remove_condition = task_data.get("remove_condition", TaskStatusCondition(agent, task.memid))
     return init_condition, stop_condition, run_condition, remove_condition
+
+
+class Time:
+    def __init__(self):
+        self.init_time_raw = time.time()
+
+    # converts from seconds to internal tick
+    def round_time(self, t):
+        return int(TICKS_PER_SEC * t)
+
+    def get_time(self):
+        return self.round_time(time.time() - self.init_time_raw)
+
+    def get_world_hour(self):
+        # returns a fraction of a day.  0 is sunrise, .5 is sunset, 1.0 is next day
+        return (time.localtime()[3] - 8 + time.localtime()[4] / 60) / 24
+
+    def add_tick(self, ticks=1):
+        time.sleep(ticks / TICKS_PER_SEC)
